+++ qt6_gui/models/audio_buffer.h
@@ -77,7 +77,7 @@ class AudioBuffer : public QObject
     [[nodiscard]] const SampleBuffer& GetChannelBuffer(size_t aChannelIndex) const;
 
     // Get the number of samples
-    [[nodiscard]] size_t NumSamples() const { return mChannelBuffers[0]->NumSamples(); }
+    [[nodiscard]] int64_t NumSamples() const { return mChannelBuffers[0]->NumSamples(); }
 
   signals:
     /**
